requirements.txt
loguru==0.6.0
peewee==3.14.9
PyMySQL==1.0.2
pyTelegramBotAPI==4.3.1
python-telegram-bot-calendar==1.0.5
requests==2.27.1
urllib3==1.26.8
python-dotenv==0.20.0

